# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "NyoStore",
    "description" : "This is a NyoStore server. Here you can find \nout all the rest api's.\n",
    "version" : "2.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/Developmenttest/NyoStore/2.1",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://authlib.nyotek.com:3010",
    "variables" : {
      "protocol" : {
        "default" : "http",
        "enum" : [ "http", "https" ]
      },
      "port" : {
        "default" : "3010",
        "enum" : [ "3010", "3020", "3030", "3040", "3050", "3060", "3070" ]
      }
    }
  } ],
...
  "paths" : {
    "/contact" : {
      "description" : "Address service (Like contact and address)",
      "post" : {
        "tags" : [ "Contact" ],
        "summary" : "Add a new contact to the store/user",
        "operationId" : "addContact",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Key"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3060",
        "description" : "Address service (Like contact and address)"
      } ]
    },
    "/contact/{key}" : {
      "get" : {
        "tags" : [ "Contact" ],
        "summary" : "Get contact by key",
        "description" : "Returns a single contact",
        "operationId" : "getContactByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "description" : "Key of contact to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContactByKey"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3060",
        "description" : "Address service (Like contact and address)"
      } ]
    },
    "/contact/{key}/contact_person" : {
      "patch" : {
        "tags" : [ "Contact" ],
        "summary" : "Update contact person by key",
        "description" : "Update contact name",
        "operationId" : "patchContactPersonByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "description" : "Key of contact to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_1"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Success"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3060",
        "description" : "Address service (Like contact and address)"
      } ]
    },
    "/contact/{key}/phone" : {
      "patch" : {
        "tags" : [ "Contact" ],
        "summary" : "Update contact phone by key",
        "description" : "Update contact name",
        "operationId" : "patchContactPhoneByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "description" : "Key of contact to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_2"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Success"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3060",
        "description" : "Address service (Like contact and address)"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Error" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Unauthorized" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "string"
          }
        }
      },
      "Forbidden" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "object"
          }
        }
      },
      "Key" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/Key_data"
          }
        },
        "example" : {
          "data" : {
            "key" : "key"
          },
          "message" : "message",
          "status" : true
        },
        "xml" : {
          "name" : "Contact"
        }
      },
...
      "body" : {
        "type" : "object",
        "properties" : {
          "contact_person" : {
            "type" : "string",
            "description" : "Name of contact person",
            "example" : "John Doe"
          },
          "phone" : {
            "type" : "string",
            "description" : "Phone no. of contact person",
            "example" : "+0 987 654 321"
          },
          "email" : {
            "type" : "string",
            "description" : "E-mail of contact person",
            "example" : "johndoe@gmail.com"
          }
        }
      },
      "body_1" : {
        "type" : "object",
        "properties" : {
          "contact_person" : {
            "type" : "string",
            "description" : "Name of contact person",
            "example" : "John Doe"
          }
        }
      },
...
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}